Athena - clean up and refactoring
