+++ properties/dev/application.properties
@@ -2,8 +2,8 @@
 server.port = 3010
 
 logging.level.*=INFO
-logging.level.org.springframework.security=DEBUG
-logging.level.org.pac4j=DEBUG
+#logging.level.org.springframework.security=DEBUG
+#logging.level.org.pac4j=DEBUG
 athena.token.secret=sssshhhh!
 athena.token.header=Athena-Auth-Token
 athena.token.expiration=604800
