--- src/main/java/com/odysseusinc/athena/config/ArachneDatabaseConfig.java
@@ -1,52 +1,0 @@
-/*
- *
- * Copyright 2018 Odysseus Data Services, inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Company: Odysseus Data Services, Inc.
- * Product Owner/Architecture: Gregory Klebanov
- * Authors: Pavel Grafkin, Vitaly Koulakov, Maria Pozhidaeva
- * Created: April 4, 2018
- *
- */
-
-package com.odysseusinc.athena.config;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
-
-import javax.sql.DataSource;
-
-@Configuration
-@EnableConfigurationProperties(DataSourceProperties.class)
-public class ArachneDatabaseConfig {
-
-    @Bean
-    public JdbcTemplate arachneJdbcTemplate(@Qualifier("arachneDataSource") DataSource dataSource) {
-
-        return new JdbcTemplate(dataSource);
-    }
-
-    @Bean
-    @ConfigurationProperties("arachne.datasource")
-    DataSource arachneDataSource() {
-
-        return DataSourceBuilder.create().build();
-    }
-}
