+++ src/main/java/com/odysseusinc/athena/repositories/v5/ConceptRelationshipV5Repository.java
@@ -27,11 +27,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.PersistenceContext;
 import java.util.Optional;
 
 @Repository
-@PersistenceContext(unitName = "athenaEntityManagerFactory")
 public interface ConceptRelationshipV5Repository
         extends JpaRepository<ConceptRelationship, Long>, JpaSpecificationExecutor<ConceptRelationship> {
