+++ src/main/java/com/odysseusinc/athena/service/mail/EmailServiceImpl.java
@@ -46,7 +46,7 @@ import static java.lang.Boolean.FALSE;
 import static java.lang.Boolean.TRUE;
 import static java.util.Arrays.asList;
 
-@Transactional(readOnly = true)
+@Transactional
 @Service
 public class EmailServiceImpl implements EmailService {
     private static final Logger LOGGER = LoggerFactory.getLogger(EmailServiceImpl.class);
